added comments and separate components and style  code  from the dashboard screen file
